Update ESpecAnalysis.py
One mistake looking through the code (self.CutOff from a legacy code) and the calibration loading had an issue. Previous calibrations had less variables and it gave back an error. Now if the variable does not exist it gives out a warning and sets the parameter to 0. (Here with the idea of CutOff, which was a variabe to get rid off artifacts). Maybe later we can just change it to the right value, but not for now,